v1.0.8: Fix CI for Linux release
Beforehand, the Linux release failed since the GitHub CLI (for `gh upload`) required being in a repository to derive information it did not have access to now. This is fixed by moving the upload step to a place where the repository is checked out, which also simplified the workflow file.