v1.0.8: Fix CI for Linux release (#14)
Beforehand, the Linux release failed since the GitHub CLI (for `gh upload`) required being in a repository to derive information it did not have access to now (resulting in `failed to run git: fatal: not a git repository (or any of the parent directories): .git`). This is fixed by checking out the repository.